+++ ext/tensor.c
@@ -23,6 +23,7 @@
 
 
 
+zend_class_entry *tensor_arraylike_ce;
 zend_class_entry *tensor_decompositions_decomposition_ce;
 zend_class_entry *tensor_tensor_ce;
 zend_class_entry *tensor_vector_ce;
@@ -42,6 +43,7 @@ static PHP_MINIT_FUNCTION(tensor)
 {
 	REGISTER_INI_ENTRIES();
 	zephir_module_init();
+	ZEPHIR_INIT(Tensor_ArrayLike);
 	ZEPHIR_INIT(Tensor_Decompositions_Decomposition);
 	ZEPHIR_INIT(Tensor_Tensor);
 	ZEPHIR_INIT(Tensor_Vector);