+++ ext/tensor/tensor.zep.h
@@ -3,27 +3,6 @@ extern ZEPHIR_API zend_class_entry *tensor_tensor_ce;
 
 ZEPHIR_INIT_CLASS(Tensor_Tensor);
 
-#if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shape, 0, 0, IS_ARRAY, 0)
-#else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shape, 0, 0, IS_ARRAY, NULL, 0)
-#endif
-ZEND_END_ARG_INFO()
-
-#if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shapestring, 0, 0, IS_STRING, 0)
-#else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shapestring, 0, 0, IS_STRING, NULL, 0)
-#endif
-ZEND_END_ARG_INFO()
-
-#if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_size, 0, 0, IS_LONG, 0)
-#else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_size, 0, 0, IS_LONG, NULL, 0)
-#endif
-ZEND_END_ARG_INFO()
-
 ZEND_BEGIN_ARG_INFO_EX(arginfo_tensor_tensor_map, 0, 0, 1)
 	ZEND_ARG_INFO(0, callback)
 ZEND_END_ARG_INFO()
@@ -127,9 +106,9 @@ ZEND_BEGIN_ARG_INFO_EX(arginfo_tensor_tensor_clipupper, 0, 0, 1)
 ZEND_END_ARG_INFO()
 
 #if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_asarray, 0, 0, IS_ARRAY, 0)
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shapestring, 0, 0, IS_STRING, 0)
 #else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_asarray, 0, 0, IS_ARRAY, NULL, 0)
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor_shapestring, 0, 0, IS_STRING, NULL, 0)
 #endif
 ZEND_END_ARG_INFO()
 
@@ -141,9 +120,6 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_tensor___tostring, 0, 0, 
 ZEND_END_ARG_INFO()
 
 ZEPHIR_INIT_FUNCS(tensor_tensor_method_entry) {
-	PHP_ABSTRACT_ME(Tensor_Tensor, shape, arginfo_tensor_tensor_shape)
-	PHP_ABSTRACT_ME(Tensor_Tensor, shapeString, arginfo_tensor_tensor_shapestring)
-	PHP_ABSTRACT_ME(Tensor_Tensor, size, arginfo_tensor_tensor_size)
 	PHP_ABSTRACT_ME(Tensor_Tensor, map, arginfo_tensor_tensor_map)
 	PHP_ABSTRACT_ME(Tensor_Tensor, reduce, arginfo_tensor_tensor_reduce)
 	PHP_ABSTRACT_ME(Tensor_Tensor, transpose, NULL)
@@ -184,7 +160,7 @@ ZEPHIR_INIT_FUNCS(tensor_tensor_method_entry) {
 	PHP_ABSTRACT_ME(Tensor_Tensor, clipUpper, arginfo_tensor_tensor_clipupper)
 	PHP_ABSTRACT_ME(Tensor_Tensor, sign, NULL)
 	PHP_ABSTRACT_ME(Tensor_Tensor, negate, NULL)
-	PHP_ABSTRACT_ME(Tensor_Tensor, asArray, arginfo_tensor_tensor_asarray)
+	PHP_ABSTRACT_ME(Tensor_Tensor, shapeString, arginfo_tensor_tensor_shapestring)
 	PHP_ABSTRACT_ME(Tensor_Tensor, __toString, arginfo_tensor_tensor___tostring)
 	PHP_FE_END
 };