+++ src/Tensor.php
@@ -2,33 +2,8 @@
 
 namespace Tensor;
 
-use IteratorAggregate;
-use ArrayAccess;
-use Countable;
-
-interface Tensor extends ArrayAccess, IteratorAggregate, Countable
+interface Tensor extends ArrayLike
 {
-    /**
-     * Return a tuple with the dimensionality of the tensor.
-     *
-     * @return int[]
-     */
-    public function shape() : array;
-
-    /**
-     * Return the shape of the tensor as a string.
-     *
-     * @return string
-     */
-    public function shapeString() : string;
-
-    /**
-     * Return the number of elements in the tensor.
-     *
-     * @return int
-     */
-    public function size() : int;
-
     /**
      * Map a function over all of the elements in the tensor.
      *
@@ -42,7 +17,7 @@ interface Tensor extends ArrayAccess, IteratorAggregate, Countable
      *
      * @param callable $callback
      * @param mixed $initial
-     * @return int|float
+     * @return mixed
      */
     public function reduce(callable $callback, $initial = 0);
 
@@ -343,11 +318,11 @@ interface Tensor extends ArrayAccess, IteratorAggregate, Countable
     public function negate();
 
     /**
-     * Return the elements of the tensor in an array.
+     * Return the shape of the tensor as a string.
      *
-     * @return array
+     * @return string
      */
-    public function asArray() : array;
+    public function shapeString() : string;
 
     /**
      * Convert the tensor into a string representation.