+++ tests/ColumnVectorTest.php
@@ -5,6 +5,7 @@ namespace Tensor\Tests;
 use Tensor\Tensor;
 use Tensor\Vector;
 use Tensor\Matrix;
+use Tensor\ArrayLike;
 use Tensor\ColumnVector;
 use PHPUnit\Framework\TestCase;
 
@@ -34,6 +35,7 @@ class ColumnVectorTest extends TestCase
         $this->assertInstanceOf(ColumnVector::class, $this->a);
         $this->assertInstanceOf(Vector::class, $this->a);
         $this->assertInstanceOf(Tensor::class, $this->a);
+        $this->assertInstanceOf(ArrayLike::class, $this->a);
     }
 
     public function test_shape()